renamed __init__.py as it was not acting as an init file
